+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this package will be documented in this file.
 
 - Created `getMeridiem` function
+- Created `getDayOfYear` function
 
 ## 0.4.1
 - Changed return of add functions